Creating skeleton for dashboard
