+++ frontend/src/components/assets/Dashboard.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Dashboard({ voterInfo, contests }) {
+	return <div>Dashboard</div>;
+}
+
+export default Dashboard;
